requirement.txt
openai
langchain
huggingface_hub
python-dotenv
streamlit